% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addranks.R
\name{addranks}
\alias{addranks}
\title{Adds rank information to a \code{coh} object}
\usage{
addranks(cohobj)
}
\arguments{
\item{cohobj}{An object of class \code{coh}}
}
\value{
\code{addranks} returns another \code{coh} object with ranks slotted now included.
If  \code{cohobj$ranks} was not NA, the \code{cohobj} is returned as is.
}
\description{
When a \code{coh} object is created, the \code{ranks} slot is NA. This function fills it in.
}
\note{
Internal function, no error checking performed
}
\examples{
#Not written yet but need some
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
