% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandtest.R
\name{bandtest}
\alias{bandtest}
\alias{bandtest.default}
\alias{bandtest.coh}
\alias{bandtest.wlmtest}
\title{Aggregate significance across a timescale band}
\usage{
bandtest(object, ...)

\method{bandtest}{default}(object, ...)

\method{bandtest}{coh}(object, band)

\method{bandtest}{wlmtest}(object, band)
}
\arguments{
\item{band}{A length-two numeric vector indicating a timescale band}

\item{obj}{An object of class \code{coh} or \code{wlmtest}, must have a non-\code{NA} 
\code{signif} slot}
}
\value{
\code{bandtest} returns an object of the same class as its first input but with a
\code{bandp} object added. Or if there was already a \code{bandp} object, the output has a 
\code{bandp} object with an additional row. For a \code{coh} object, the \code{bandp} object 
is a data frame with four columns, the first two indicating the timescale band and the third 
an associated p-value for the test of coherence over that band. The fourth column is the 
average phase over the band. For a \code{wlmtest} object, the result is only the first three
of the above columns.
}
\description{
Computes the aggregate significance of coherence or of a wavelet linear model test object
across a timescale band, accounting for non-independence of timescales. Also gets the 
average phase across the band, in the case of coherence.
}
\references{
Sheppard, L.W., et al. (2016) Changes in large-scale climate alter spatial synchrony of aphid 
pests. Nature Climate Change. DOI: 10.1038/nclimate2881
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
