% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohbandtest.R
\name{cohbandtest}
\alias{cohbandtest}
\alias{cohbandtest.default}
\alias{cohbandtest.coh}
\title{Aggregate significance of coherence across a timescale band}
\usage{
cohbandtest(object, ...)

\method{cohbandtest}{default}(object, ...)

\method{cohbandtest}{coh}(cohobj, band)
}
\arguments{
\item{cohobj}{An object of class \code{coh}, must have a non-\code{NA} \code{signif} slot}

\item{band}{A length-two numeric vector indicating a timescale band}
}
\value{
\code{cohbandtest} returns an object of the same class as its first input but with a
\code{bandp} object added. Or if there was already a \code{bandp} object, the output has a 
\code{bandp} object with an additional row. The \code{bandp} object is a data frame with four
columns, the first two indicating the timescale band and the third an associated p-value for
the test of coherence over that band. The fourth column is the average phase over the band.
}
\description{
Computes the aggregate significance of coherence across a timescale band, accounting for
non-independence of timescales. Also gets the average phase across the band.
}
\references{
Sheppard, L.W., et al. (2016) Changes in large-scale climate alter spatial synchrony of aphid 
pests. Nature Climate Change. DOI: 10.1038/nclimate2881
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
