% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_stdat.R
\name{errcheck_stdat}
\alias{errcheck_stdat}
\title{Error check for appropriate spatio-temporal data}
\usage{
errcheck_stdat(times, dat, callfunc)
}
\arguments{
\item{times}{the times of measurement, spacing 1}

\item{dat}{each row is a time series - must have at least two rows}

\item{callfunc}{the function calling this one, for error tracking}
}
\value{
\code{errcheck_stdat} returns nothing but throws and error if inputs not appropriate
}
\description{
Error checking whether a times vector and a matrix with each row a time
series make a legitimate spatio-temporal data set for wavelet analysis
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
