% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_times.R
\name{errcheck_times}
\alias{errcheck_times}
\title{Error check \code{times}}
\usage{
errcheck_times(times, callfunc)
}
\arguments{
\item{times}{Tests whether this is a numeric vector with unit-spaced increasing values}

\item{callfunc}{Function calling this one, for better error messaging}
}
\value{
\code{errcheck_times} returns nothing but throws and error if the conditions are not met
}
\description{
Error check whether a vector can represent times at which data suitable 
for wavelet transforms were measured
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
