% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_tts.R
\name{errcheck_tts}
\alias{errcheck_tts}
\title{Error check whether inputs are suitable for a tts object}
\usage{
errcheck_tts(times, timescales, values, callfunc)
}
\arguments{
\item{times}{times of measurement}

\item{values}{a times by timescales matrix}

\item{callfunc}{the function from which this one was called, for error tracking}

\item{timescale}{timescales of analysis}
}
\value{
\code{errcheck_tts} returns nothing but throws and error if inputs not appropriate
}
\description{
Error check whether inputs are suitable for a tts object
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
