% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnphase.R
\name{mnphase}
\alias{mnphase}
\title{Mean phase of coherence}
\usage{
mnphase(nums)
}
\arguments{
\item{nums}{A vector of complex numbers}
}
\value{
\code{mnphase} returns the mean phase
}
\description{
Gets the mean phaseof a bunch of complex numbers
}
\note{
Internal funcion, no error catching
}
\examples{
#Not written yet but need some
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
