% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normforcoh.R
\name{normforcoh}
\alias{normforcoh}
\title{Normalization for the \code{coh} function}
\usage{
normforcoh(W, norm)
}
\arguments{
\item{W}{An array of wavelet transforms, locations by times by timescales}

\item{norm}{The normalization of wavelet transforms to use. Controls the version of the coherence that is performed. One of "none", "phase", "powall", "powind". See details section of the documentation for \code{coh}.}
}
\value{
\code{normforcoh} returns an array the same dimensions as W of normalized transforms
}
\description{
A convenience function for performing the normalization step for the \code{coh} function.
}
\note{
Internal function, no error checking
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
