% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.R
\name{plotmap}
\alias{plotmap}
\title{Map clusters from a \code{clust} object}
\usage{
plotmap(inclust, spltlvl = length(inclust$clusters), nodesize = c(1, 3),
  filename = NA)
}
\arguments{
\item{inclust}{A \code{clust} object, as created with \code{wsyn::clust}}

\item{spltlvl}{The split level in the clustering to use. This is the index of inclust$clusters.
Default the final split.}

\item{nodesize}{A length = 2 vector giving the minimum and maximum node size for plotting. Defaults to c(1,3).}

\item{filename}{a filename, possibly including path info, but without a file extension. If present,
exports the plot as a .pdf using the specified filename. Default \code{NA} uses the default plotting
device.}
}
\value{
\code{plotmap} produces a map.
}
\description{
Produces a map of the locations of sampling for a \code{clust} object, with colors indicating 
module (cluster) identity.
The sizes of nodes (locations) are scaled according to the strength of membership in its module.
}
