% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotphase.R
\name{plotphase}
\alias{plotphase}
\alias{plotphase.coh}
\alias{plotphase.default}
\title{For plotting the phases of values in \code{tts} and \code{coh} objects}
\usage{
plotphase(object, ...)

\method{plotphase}{coh}()

\method{plotphase}{default}(object, ...)
}
\description{
For plotting the phases of values in \code{tts} objects (and derived classes) 
against time and timescale, and \code{coh} objects against timescale
}
\examples{
#Not written yet but need some

}
\references{
Sheppard, L.W., et al. (2016) Changes in large-scale climate alter spatial synchrony of aphid 
pests. Nature Climate Change. DOI: 10.1038/nclimate2881
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
