% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotwpmfs.R
\name{plotwpmfs}
\alias{plotwpmfs}
\alias{plotwpmfs.clust}
\alias{plotwpmfs.default}
\title{Plots wavelet phasor means fields of clusters in a \code{clust} object}
\usage{
plotwpmfs(object, ...)

\method{plotwpmfs}{clust}(object, spltlev = length(object$clusters),
  filename = NA)

\method{plotwpmfs}{default}(object, ...)
}
\arguments{
\item{object}{A \code{clust} object}

\item{filename}{The filename (without extension) to use for saving plots. One plot for each cluster,
names (if given) formed by appending the cluster number to the end of filename and then adding an 
extension. Default NA uses the defulat plotting device.}

\item{spltlvl}{The level of split to use. Must be between 1 and length(object$clusters). The former 
corresponds to all locations (prior to any splitting) and the latter to the final set of clusters.}
}
\value{
\code{plotwpmfs} returns nothing but generates plots.
}
\description{
Plots wavelet phasor means fields of clusters in a \code{clust} object
}
\examples{
#Not written yet but need some
 
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
