% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power}
\alias{power}
\alias{power.tts}
\title{Power of a \code{tts} object}
\usage{
power(object, ...)

\method{power}{tts}(object, ploton = TRUE)
}
\arguments{
\item{object}{A \code{tts} object}
}
\value{
\code{power} returns a data frame with columns timescales and power
}
\description{
Returns the power of a \code{tts} object, i.e., the mean over
time of the square magnitude (which is a function of timescale)
}
\examples{
Do not have any yet but need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
