% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predsync.R
\name{predsync}
\alias{predsync}
\alias{predsync.wlm}
\title{Predicted synchrony of a wavelet linear model}
\usage{
predsync(object, ...)

\method{predsync}{wlm}(wlmobj)
}
\arguments{
\item{wlmobj}{A \code{wlm} object}
}
\value{
\code{predsync} returns a \code{tts} object. Plotting the magnitude
(see \code{plotmag}) displays a picture of predicted synchrony versus time and
timescale that is comparable with the wavelet mean field (see \code{wmf}) of
the response variable of the model. Calling the \code{power} function on that 
\code{tts} object should give the same results as one of the columns of output
of \code{syncexpl}. Only \code{norm="powall"} implemented so far.
}
\description{
Predicted synchrony of a \code{wlm} object. This is described in the
first paragraph of Appendix S15 of Sheppard et al (2018).
}
\examples{
#Not written yet but need some

}
\references{
Sheppard, LW et al. (2018) Synchrony is more than its top-down and climatic parts: interacting 
Moran effects on phytoplankton in British seas, In review.
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
