% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_wsyn.R
\name{print.summary_wsyn}
\alias{print.summary_wsyn}
\title{Print method for \code{summary.wsyn} class}
\usage{
print.summary_wsyn(x, ...)
}
\arguments{
\item{x}{A \code{summary.wsyn} object}
}
\value{
\code{print.summary_wsyn} is called for its effect of
printing to the screen.
}
\description{
Print method for \code{summary.wsyn} class
}
\examples{
times<-1:10
timescales<-1/c(1:10)
values<-matrix(1,length(times),length(timescales))
h<-tts(times,timescales,values)
print(summary(h))

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
