% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setmints.R
\name{setmints}
\alias{setmints}
\title{Shifts a vector according to the argument mints}
\usage{
setmints(ts, mints)
}
\arguments{
\item{ts}{A vector of numeric values representing a time series}

\item{mints}{The time series is shifted to have this minimum value. Default NA means use the smallest difference 
between consecutive, distinct sorted values of the time series. NaN means perform no shift.}
}
\value{
\code{setmints} returns the shifted vector.

Daniel Reuman, \email{reuman@ku.edu}
}
\description{
Shifts a vector according to the argument mints
}
\note{
This is an internal function, and no error checking is done.
}
