% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syncexpl.R
\name{syncexpl}
\alias{syncexpl}
\alias{syncexpl.wlm}
\title{Amount of synchrony explained, and related quantities}
\usage{
syncexpl(object)

\method{syncexpl}{wlm}(object)
}
\arguments{
\item{object}{A \code{wlm} object}
}
\value{
\code{syncexpl} returns a data frame with columns for \code{timescales},
\code{sync} (the time-averaged square magnitude of the wavelet mean field of the
responbse transforms), \code{syncexpl} (synchrony explained by the model 
predictors), columns named for each predictor (synchrony explained by that 
predictor), \code{interactions} (synchrony explained by all interaction effects),
columns named for each pair of predictors (synchrony explained by individual
pairwise interactions). There are also columns for \code{crossterms} and 
\code{resids} (residuals) that must be small for a given timescale band for
the other results to be meaningful. All columns are functions of timescales.
}
\description{
Gives amount of synchrony explained by a wavelet linear model, as a function of 
timescale, and related quantities (see details)
}
\details{
This function only works for \code{norm="powall"} at present. See
Sheppard et al (2018) for details of the meaning and computation of the 
columns.
}
\examples{
#Not written yet but need some
 
}
\references{
Sheppard, LW et al. (2018) Synchrony is more than its top-down and climatic parts: interacting 
Moran effects on phytoplankton in British seas, In review.
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
