% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addranks.R
\name{addranks}
\alias{addranks}
\title{Adds rank information to a \code{coh} or \code{wlmtest} object}
\usage{
addranks(obj)
}
\arguments{
\item{obj}{An object of class \code{coh} or \code{wlmtest}}
}
\value{
\code{addranks} returns another \code{coh} or \code{wlmtest} object with ranks 
slot now included. If  \code{obj$ranks} was not NA, the object is returned as is.
}
\description{
When a \code{coh} or \code{wlmtets} object is created, the \code{ranks} slot is NA. 
This function fills it in.
}
\note{
Internal function, no error checking performed
}
\seealso{
\code{\link{coh}}, \code{\link{wlmtest}}, \code{\link{bandtest}}, \code{browseVignettes("wsyn")}
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
