% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addwmfs.R
\name{addwmfs}
\alias{addwmfs}
\title{Adds wavelet mean field information to a \code{clust} object}
\usage{
addwmfs(obj)
}
\arguments{
\item{obj}{An object of class \code{clust}}
}
\value{
\code{addwmfs} returns another \code{clust} object with \code{wmfs} slot now included. 
If  \code{obj$wmfs} was not NA, the object is returned as is.
}
\description{
When a \code{clust} object is created, the \code{wmfs} slot is NA. This function fills it in.
}
\note{
Internal function, no error checking performed
}
\examples{
#Not written yet but need some
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
