% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addwpmfs.R
\name{addwpmfs}
\alias{addwpmfs}
\title{Adds wavelet phasor mean field information to a \code{clust} object}
\usage{
addwpmfs(obj, level = 1:length(obj$clusters), sigmethod = "quick",
  nrand = 1000)
}
\arguments{
\item{obj}{An object of class \code{clust}}

\item{level}{The clustering level(s) to use. 1 corresponds to no clustering. The default is all 
levels of clustering.}

\item{sigmethod}{Method for significance testing the \code{wpmf}, one of \code{quick}, \code{fft}, 
\code{aaft} (see details of the \code{wpmf} function)}

\item{nrand}{The number of randomizations to be used for significance testing}
}
\value{
\code{addwpmfs} returns another \code{clust} object with \code{wpmfs} slot now included,
or more filled in than it was previously. With values of \code{sigmethod} other than 
\code{"quick"}, this function can be slow, particularly with large \code{nrand}. So in that
case the user may want to set \code{level} equal only to one clustering level of interest.
Unlike \code{wmf}, old values in \code{obj$wpmfs} are overwritten.
}
\description{
When a \code{clust} object is created, the \code{wpmfs} slot is NA. This function fills it in, 
or adds to it.
}
\details{
This function uses the values of \code{scale.min}, \code{scale.max.input}, 
\code{sigma} and \code{f0} stored in \code{obj$methodspecs}. It is possible to create 
a clust object with bad values for these slots. This function throws an error in that 
case. You can use a correlation-based method for calculating the synchrony matrix and 
still pass values of \code{scale.min}, \code{scale.max.input}, \code{sigma} and \code{f0} 
to \code{clust} (in fact, this happens by default) - they won't be used by \code{clust}, 
but they will be there for later use by \code{addwmfs} and \code{addwpmfs}.
}
\examples{
#Not written yet but need some

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
