% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotphase.R
\name{plotphase}
\alias{plotphase}
\alias{plotphase.coh}
\alias{plotphase.default}
\title{For plotting the phases of values in \code{tts} and \code{coh} objects}
\usage{
plotphase(object, ...)

\method{plotphase}{coh}(object, bandprows = "all", filename = NA, ...)

\method{plotphase}{default}(object, ...)
}
\arguments{
\item{object}{A \code{coh} object.}

\item{...}{Passed from the generic to specific methods. Not currently used.}

\item{bandprows}{The rows of \code{object$bandp} for which to display p-value results in the plot}

\item{filename}{Filename (without extension), for saving as pdf. Default value NA saves no file and uses the default graphics device.}
}
\description{
For plotting the phases of values in \code{tts} objects (and derived classes) 
against time and timescale, and \code{coh} objects against timescale
}
\examples{
#Not written yet but need some

}
\references{
Sheppard, L.W., et al. (2016) Changes in large-scale climate alter spatial synchrony of aphid 
pests. Nature Climate Change. DOI: 10.1038/nclimate2881
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
