% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotrank.R
\name{plotrank}
\alias{plotrank}
\alias{plotrank.coh}
\alias{plotrank.default}
\title{Plots \code{ranks} slot for \code{coh} objects}
\usage{
plotrank(object, ...)

\method{plotrank}{coh}(object, sigthresh = 0.95, bandprows = "all",
  filename = NA)

\method{plotrank}{default}(object, ...)
}
\arguments{
\item{object}{A \code{coh} object. Must have a non-\code{NA} \code{signif} slot.}

\item{sigthresh}{Significance threshold(s). Numeric vector with values between 0 and 1. Typically 0.95, 0.99, 0.999, etc. The threshhold(s) are plotted on the rank plot as dashed horizontal line(s).}

\item{bandprows}{The rows of \code{object$bandp} for which to display p-value results in the plot}

\item{filename}{Filename (without extension), for saving as pdf. Default value NA saves no file and uses the default graphics device.}
}
\description{
Plots the \code{ranks} slot for \code{coh} objects to help identify statistical significance of coherence
}
\details{
The plot shows \code{object$ranks$coher} versus \code{log(1/object$timescales)}.
Horizontal axis ticks are labeled as timescales, but are spaced on the axis as 
log(1/timescale), i.e., log frequencies. p-values from \code{object$bandp} are displayed
above the rank plot.
}
\examples{
#Not written yet but need some

}
\references{
Sheppard, L.W., et al. (2016) Changes in large-scale climate alter spatial synchrony of aphid 
pests. Nature Climate Change. DOI: 10.1038/nclimate2881
}
\author{
Thomas Anderson, \email{anderstl@gmail.com}, Jon Walter, \email{jaw3es@virginia.edu}; Lawrence 
Sheppard, \email{lwsheppard@ku.edu}; Daniel Reuman, \email{reuman@ku.edu}
}
