% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_wsyn.R
\name{print.summary_wsyn}
\alias{print.summary_wsyn}
\title{Print method for \code{summary.wsyn} class}
\usage{
\method{print}{summary_wsyn}(x, ...)
}
\arguments{
\item{x}{A \code{summary.wsyn} object}

\item{...}{Not currently used. Included for argument consistency
with existing generics.}
}
\value{
\code{print.summary_wsyn} is called for its effect of
printing to the screen.
}
\description{
Print method for \code{summary.wsyn} class
}
\examples{
times<-1:10
timescales<-1/c(1:10)
values<-matrix(1,length(times),length(timescales))
h<-tts(times,timescales,values)
print(summary(h))

}
\seealso{
\code{\link{tts_methods}}, \code{\link{wt_methods}}, \code{\link{wmf_methods}}, \code{\link{wpmf_methods}}, 
\code{\link{coh_methods}}, \code{\link{wlm_methods}}, \code{\link{wlmtest_methods}}, \code{\link{clust_methods}},
\code{browseVignettes("wsyn")}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
