% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tts_methods.R, R/wpmf_methods.R, R/wt_methods.R
\name{setget_methods}
\alias{setget_methods}
\alias{set_times}
\alias{set_times.default}
\alias{set_timescales}
\alias{set_timescales.default}
\alias{set_values}
\alias{set_values.default}
\alias{get_times}
\alias{get_times.default}
\alias{get_timescales}
\alias{get_timescales.default}
\alias{get_values}
\alias{get_values.default}
\alias{set_signif}
\alias{set_signif.default}
\alias{get_signif}
\alias{get_signif.default}
\alias{set_dat}
\alias{set_dat.default}
\alias{set_wtopt}
\alias{set_wtopt.default}
\alias{get_dat}
\alias{get_dat.default}
\alias{get_wtopt}
\alias{get_wtopt.default}
\title{Set and get methods for classes in the \code{wsyn} package}
\usage{
set_times(obj, newval)

\method{set_times}{default}(obj, newval)

set_timescales(obj, newval)

\method{set_timescales}{default}(obj, newval)

set_values(obj, newval)

\method{set_values}{default}(obj, newval)

get_times(obj)

\method{get_times}{default}(obj)

get_timescales(obj)

\method{get_timescales}{default}(obj)

get_values(obj)

\method{get_values}{default}(obj)

set_signif(obj, newval)

\method{set_signif}{default}(obj, newval)

get_signif(obj)

\method{get_signif}{default}(obj)

set_dat(obj, newval)

\method{set_dat}{default}(obj, newval)

set_wtopt(obj, newval)

\method{set_wtopt}{default}(obj, newval)

get_dat(obj)

\method{get_dat}{default}(obj)

get_wtopt(obj)

\method{get_wtopt}{default}(obj)
}
\arguments{
\item{obj}{An object of one of the classes defined in the package}

\item{newval}{A newvalue of the slot in question, for the \code{set_*} methods}
}
\value{
\code{set_*} methods throw an error - setting of individual
slots is not allowed, as it breaks consistency with the other slots.
\code{get_*} just returns the value in question.
}
\description{
Set and get methods for classes in the \code{wsyn} package. There
are methods for each slot of each class, named \code{set_*} and
\code{get_*} for \code{*} the slot name. Below are listed function
specs for the generics and the default methods.
}
\details{
There are methods for the \code{tts}, \code{wt}, \code{wmf},
\code{wpmf}, \code{coh}, \code{wlm}, \code{wlmtest}, and \code{clust}
classes.
}
\examples{
times<-1:10
timescales<-1/c(1:10)
values<-matrix(1,length(times),length(timescales))
h<-tts(times,timescales,values)
get_times(h)

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
