% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_methods.R, R/coh_methods.R,
%   R/tts_methods.R, R/wlm_methods.R, R/wlmtest_methods.R, R/wpmf_methods.R,
%   R/wt_methods.R
\name{set_adj}
\alias{set_adj}
\alias{set_adj.default}
\alias{set_clusters}
\alias{set_clusters.default}
\alias{set_modres}
\alias{set_modres.default}
\alias{set_mns}
\alias{set_mns.default}
\alias{set_coords}
\alias{set_coords.default}
\alias{set_methodspecs}
\alias{set_methodspecs.default}
\alias{set_wmfs}
\alias{set_wmfs.default}
\alias{set_wpmfs}
\alias{set_wpmfs.default}
\alias{get_adj}
\alias{get_adj.default}
\alias{get_clusters}
\alias{get_clusters.default}
\alias{get_modres}
\alias{get_modres.default}
\alias{get_mns}
\alias{get_mns.default}
\alias{get_coords}
\alias{get_coords.default}
\alias{get_methodspec}
\alias{get_methodspec.default}
\alias{get_wmfs}
\alias{get_wmfs.default}
\alias{get_wpmfs}
\alias{get_wpmfs.default}
\alias{set_coher}
\alias{set_coher.default}
\alias{set_dat1}
\alias{set_dat1.default}
\alias{set_dat2}
\alias{set_dat2.default}
\alias{set_norm}
\alias{set_norm.default}
\alias{set_sigmethod}
\alias{set_sigmethod.default}
\alias{set_ranks}
\alias{set_ranks.default}
\alias{set_bandp}
\alias{set_bandp.default}
\alias{get_coher}
\alias{get_coher.default}
\alias{get_dat1}
\alias{get_dat1.default}
\alias{get_dat2}
\alias{get_dat2.default}
\alias{get_norm}
\alias{get_norm.default}
\alias{get_sigmethod}
\alias{get_sigmethod.default}
\alias{get_ranks}
\alias{get_ranks.default}
\alias{get_bandp}
\alias{get_bandp.default}
\alias{setget_methods}
\alias{set_times}
\alias{set_times.default}
\alias{set_timescales}
\alias{set_timescales.default}
\alias{set_values}
\alias{set_values.default}
\alias{get_times}
\alias{get_times.default}
\alias{get_timescales}
\alias{get_timescales.default}
\alias{get_values}
\alias{get_values.default}
\alias{set_coefs}
\alias{set_coefs.default}
\alias{set_modval}
\alias{set_modval.default}
\alias{set_wts}
\alias{set_wts.default}
\alias{get_coefs}
\alias{get_coefs.default}
\alias{get_modval}
\alias{get_modval.default}
\alias{get_wts}
\alias{get_wts.default}
\alias{set_wlmobj}
\alias{set_wlmobj.default}
\alias{set_drop}
\alias{set_drop.default}
\alias{get_wlmobj}
\alias{get_wlmobj.default}
\alias{get_drop}
\alias{get_drop.default}
\alias{set_signif}
\alias{set_signif.default}
\alias{get_signif}
\alias{get_signif.default}
\alias{set_dat}
\alias{set_dat.default}
\alias{set_wtopt}
\alias{set_wtopt.default}
\alias{get_dat}
\alias{get_dat.default}
\alias{get_wtopt}
\alias{get_wtopt.default}
\title{Set and get methods for classes in the \code{wsyn} package}
\usage{
set_adj(obj, newval)

\method{set_adj}{default}(obj, newval)

set_clusters(obj, newval)

\method{set_clusters}{default}(obj, newval)

set_modres(obj, newval)

\method{set_modres}{default}(obj, newval)

set_mns(obj, newval)

\method{set_mns}{default}(obj, newval)

set_coords(obj, newval)

\method{set_coords}{default}(obj, newval)

set_methodspecs(obj, newval)

\method{set_methodspecs}{default}(obj, newval)

set_wmfs(obj, newval)

\method{set_wmfs}{default}(obj, newval)

set_wpmfs(obj, newval)

\method{set_wpmfs}{default}(obj, newval)

get_adj(obj)

\method{get_adj}{default}(obj)

get_clusters(obj)

\method{get_clusters}{default}(obj)

get_modres(obj)

\method{get_modres}{default}(obj)

get_mns(obj)

\method{get_mns}{default}(obj)

get_coords(obj)

\method{get_coords}{default}(obj)

get_methodspec(obj)

\method{get_methodspec}{default}(obj)

get_wmfs(obj)

\method{get_wmfs}{default}(obj)

get_wpmfs(obj)

\method{get_wpmfs}{default}(obj)

set_coher(obj, newval)

\method{set_coher}{default}(obj, newval)

set_dat1(obj, newval)

\method{set_dat1}{default}(obj, newval)

set_dat2(obj, newval)

\method{set_dat2}{default}(obj, newval)

set_norm(obj, newval)

\method{set_norm}{default}(obj, newval)

set_sigmethod(obj, newval)

\method{set_sigmethod}{default}(obj, newval)

set_ranks(obj, newval)

\method{set_ranks}{default}(obj, newval)

set_bandp(obj, newval)

\method{set_bandp}{default}(obj, newval)

get_coher(obj)

\method{get_coher}{default}(obj)

get_dat1(obj)

\method{get_dat1}{default}(obj)

get_dat2(obj)

\method{get_dat2}{default}(obj)

get_norm(obj)

\method{get_norm}{default}(obj)

get_sigmethod(obj)

\method{get_sigmethod}{default}(obj)

get_ranks(obj)

\method{get_ranks}{default}(obj)

get_bandp(obj)

\method{get_bandp}{default}(obj)

set_times(obj, newval)

\method{set_times}{default}(obj, newval)

set_timescales(obj, newval)

\method{set_timescales}{default}(obj, newval)

set_values(obj, newval)

\method{set_values}{default}(obj, newval)

get_times(obj)

\method{get_times}{default}(obj)

get_timescales(obj)

\method{get_timescales}{default}(obj)

get_values(obj)

\method{get_values}{default}(obj)

set_coefs(obj, newval)

\method{set_coefs}{default}(obj, newval)

set_modval(obj, newval)

\method{set_modval}{default}(obj, newval)

set_wts(obj, newval)

\method{set_wts}{default}(obj, newval)

get_coefs(obj)

\method{get_coefs}{default}(obj)

get_modval(obj)

\method{get_modval}{default}(obj)

get_wts(obj)

\method{get_wts}{default}(obj)

set_wlmobj(obj, newval)

\method{set_wlmobj}{default}(obj, newval)

set_drop(obj, newval)

\method{set_drop}{default}(obj, newval)

get_wlmobj(obj)

\method{get_wlmobj}{default}(obj)

get_drop(obj)

\method{get_drop}{default}(obj)

set_signif(obj, newval)

\method{set_signif}{default}(obj, newval)

get_signif(obj)

\method{get_signif}{default}(obj)

set_dat(obj, newval)

\method{set_dat}{default}(obj, newval)

set_wtopt(obj, newval)

\method{set_wtopt}{default}(obj, newval)

get_dat(obj)

\method{get_dat}{default}(obj)

get_wtopt(obj)

\method{get_wtopt}{default}(obj)
}
\arguments{
\item{obj}{An object of one of the classes defined in the package}

\item{newval}{A newvalue of the slot in question, for the \code{set_*} methods}
}
\value{
\code{set_*} methods throw an error - setting of individual
slots is not allowed, as it breaks consistency with the other slots.
\code{get_*} just returns the value in question.
}
\description{
Set and get methods for classes in the \code{wsyn} package. There
are methods for each slot of each class, named \code{set_*} and
\code{get_*} for \code{*} the slot name. Below are listed function
specs for the generics and the default methods.
}
\details{
There are methods for the \code{tts}, \code{wt}, \code{wmf},
\code{wpmf}, \code{coh}, \code{wlm}, \code{wlmtest}, and \code{clust}
classes. See documentation for the generator functions for these classes
(which in all cases have the same name as the class) for lists of slots
for each class.
}
\examples{
times<-1:10
timescales<-1/c(1:10)
values<-matrix(1,length(times),length(timescales))
h<-tts(times,timescales,values)
get_times(h)

}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
